refactor: 타이머 상수 및 로직 분리
