+++ src/basic/constants.js
@@ -27,3 +27,8 @@ export const PRODUCT_DEFAULT_DISCOUNT_RATES = {
   LAPTOP_POUCH: 0.05,
   SPEAKER: 0.25,
 };
+
+export const TIMER_INTERVAL = {
+  LIGHTNING: 30000,
+  RECOMMEND: 60000,
+};
